split up tarnation grammar into individual modules
